+++ memflow-win32/src/kernel/start_block/x64.rs
@@ -34,7 +34,7 @@ pub fn find_lowstub(stub: &[u8]) -> Result<StartBlock> {
 
 fn find_pt(addr: Address, mem: &[u8]) -> Option<Address> {
     // TODO: global define / config setting
-    let max_mem = size::gb(512) as u64;
+    let max_mem = size::gb(512);
 
     let pte = u64::from_le_bytes(mem[0..8].try_into().unwrap());
